requirements.txt
asgiref==3.5.2
crispy-bootstrap5==0.7
Django==4.0.6
django-countries==7.4.2
django-crispy-forms==1.14.0
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==9.2.0
pytz==2022.6
sqlparse==0.4.2
typing_extensions==4.4.0
